gettext: use LOCALEDIR as an alternative for finding the locale dir
The variable LOCALEDIR can be set in Makefile.conf. At first, I wanted
to use the 'portable' ('locals') but unfortunately, by default,
the directory searched when using a portable installation is

    <prefix>/locale

but Homebrew expects locale things to be installed in the 'share'
dir in the prefix:

    <prefix>/share/locale

Thanks Alexis for your help!