doc/Makefile: fix a bug on the osx version of gnumake-3.81
With the osx version of gnumake-3.81 (provided by the command line
developer tools, installed by 'xcode-select --install'), it seems that
the recipe command
                sed -i -e '/setuplocale/a \ <-- ISSUE HERE
    \\usepackage{xeCJK}\
    \\setCJKmainfont{IPAexMincho}\
    \\setCJKsansfont{IPAexGothic}\
    \\setCJKmonofont{IPAexGothic}\
    ' $<
gives the wrong auto-multiple-choice.ja.tex (notice the missing
backslash before 'usepackage{xeCJK}'):
    \setuplocale{ja}
    usepackage{xeCJK}
    \setCJKmainfont{IPAexMincho}
    \setCJKsansfont{IPAexGothic}
    \setCJKmonofont{IPAexGothic}
I downloaded and compiled gnumake 3.81, 3.81 and 4.0 and they all
yield the right result (the backslash is not missing). As a workaround,
I propose to add a second backslash preceeded by a space, which
mitigates the issue.