pdftk-alternative: AMC-read-pdfform: use our own command instead of pdftk
The rationale is that we know that at some point, 'pdftk' will be
retired as it is not actively maintained and Ubuntu will soon drop
it from its repositories: https://bugs.launchpad.net/bugs/1757314.

This intent in mind, we wrote a C program in pdfformfields.c that mimics
pdftk's dump_data_fields_utf8 command. It relies on the poppler library
(which is well maintained).