pdftk-alternative: AMC-getimages: add 'qpdf' for splitting pdfs in AMC-getimages
AMC-getimages tries to run 'pdfimages' (poppler), then tries qpdf and
finally pdftk.

It is preferable that 'pdfimages' is picked before 'qpdf' because many
scanners produce a PDF scan which is a simple container with the actual
raster images of the pages inside. 'pdfimages' extracts those images
without any modification, so that both operations "split pages" and
"convert to raster" are done with little cost. If we use 'qpdf' for
splitting instead of pdfimages, AMC will then convert the one-page PDF
files to raster, which will take time and may introduce a quality loss.