Substitute LaTeX pkg bophook w/ atbegshi or hooks
bophook is a legacy package; current versions of LaTeX use hooks
to achieve the same in a standard fashion, and the "traditional"
atbegshi interface is so popular that it is emulated by the newer
kernels. Since bophook has issues with (at least) hyperref, we
replace it with these newer approaches.